remove saved type nbt and move methods
