+++ src/main/java/com/cleanroommc/modularui/api/IItemStackLong.java
@@ -11,25 +11,6 @@ import net.minecraft.nbt.NBTTagCompound;
 
 public interface IItemStackLong {
 
-    public static @Nullable IItemStackLong loadItemStackFromNBT(@Nonnull NBTTagCompound nbt) {
-        int type = nbt.getInteger("type");
-        return switch(type) {
-            case 1 -> ItemStackLong.loadFromNBT(nbt);
-            default -> null;
-        };
-    }
-
-    public static boolean areItemStacksEqual(@Nullable IItemStackLong a, @Nullable IItemStackLong b) {
-        if (a == null && b == null) return true;
-        if (a == null || b == null) return false;
-        if (a.getItem() != b.getItem()) return false;
-        if (a.getItemDamage() != b.getItemDamage()) return false;
-        if (a.getStackSize() != b.getStackSize()) return false;
-        if (a.getTagCompound() == null && b.getTagCompound() == null) return true;
-        if (a.getTagCompound() == null || b.getTagCompound() == null) return false;
-        return a.getTagCompound().equals(b.getTagCompound());
-    }
-
     /**
      * @return The stacksize the item has currently
      */
