+++ src/main/java/com/cleanroommc/modularui/utils/item/ItemStackLong.java
@@ -117,7 +117,6 @@ public class ItemStackLong implements IItemStackLong {
 
     @Override
     public @Nonnull NBTTagCompound writeToNBT(@Nonnull NBTTagCompound nbt) {
-        nbt.setInteger("type", 1);
         ItemStack itemStack = getAsItemStack();
         itemStack.setTagCompound(getTagCompound());
         itemStack.writeToNBT(nbt);
@@ -131,4 +130,16 @@ public class ItemStackLong implements IItemStackLong {
         ItemStackLong item = new ItemStackLong(itemStack.getItem(), nbt.getLong("maxStackSizeLong"), itemStack.getItemDamage(), nbt.getLong("stackSizeLong"), itemStack.getTagCompound());
         return item;
     }
+
+    public static boolean areItemStacksEqual(@Nullable IItemStackLong a, @Nullable IItemStackLong b) {
+        if (a == null && b == null) return true;
+        if (a == null || b == null) return false;
+        if (a.getItem() != b.getItem()) return false;
+        if (a.getItemDamage() != b.getItemDamage()) return false;
+        if (a.getStackSize() != b.getStackSize()) return false;
+        if (a.getTagCompound() == null && b.getTagCompound() == null) return true;
+        if (a.getTagCompound() == null || b.getTagCompound() == null) return false;
+        return a.getTagCompound().equals(b.getTagCompound());
+    }
+
 }
